pyproject.toml
[project]
name = "lucyui"
description = "User Interface and Layout Management Library for Pygame"
version = "0.0.3"
readme = "README.md"
authors = [
    {name="Kadir Aksoy"},
]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Topic :: Games/Entertainment",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: pygame",
]
dependencies = [
    "pygame-ce",
]
requires-python = ">=3.10"

[project.urls]
homepage = "https://github.com/kadir014/lucy-ui"
bugtracker = "https://github.com/kadir014/lucy-ui/issues"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"